add api in pengaduan dashboard

+++ config/fetchApi.ts
@@ -1,25 +1,22 @@
-import axios, { AxiosInstance, AxiosResponse, AxiosError } from 'axios';
-import EncryptStorage from 'encrypt-storage';
+// src/config/fetchApi.ts
+import axios, { AxiosInstance, AxiosResponse, AxiosError } from "axios";
+import { EncryptStorage } from "encrypt-storage";
 
 interface UserInfo {
   token: string;
-  // Add other user info properties here if needed
 }
 
 const fetchApi = (type: string = ""): AxiosInstance => {
-  const encryptStorage = new EncryptStorage(process.env.REACT_APP_SECRET_KEY as string);
-  const baseURL = process.env.REACT_APP_SERVICE as string;
+  const encryptStorage = new EncryptStorage(process.env.NEXT_PUBLIC_ENCRYPT_STORAGE_SECRET_KEY!);
+  const baseURL = process.env.NEXT_PUBLIC_API_URL!;
 
-  // Get Token
-  const userInfo: UserInfo | null = encryptStorage.getItem("___info");
+  const userInfo: UserInfo | any = encryptStorage.getItem("info");
   const accessToken: string = userInfo?.token || "";
 
   const api: AxiosInstance = axios.create({
-    baseURL: baseURL,
+    baseURL,
     headers: {
       Authorization: `Bearer ${accessToken}`,
-      "Access-Control-Allow-Origin": true,
-      "Access-Control-Allow-Credentials": true,
       "Accept-Language": "en",
     },
   });
@@ -27,17 +24,12 @@ const fetchApi = (type: string = ""): AxiosInstance => {
   api.interceptors.response.use(
     (response: AxiosResponse) => response,
     (error: AxiosError) => {
-      if (type === "auth") {
-        return Promise.reject(error);
-      } else {
-        if (error?.response?.status === 403) {
-          // Token Expired
-          encryptStorage.clear();
-          localStorage.clear();
-          window.location.href = "/login";
-        }
-        return Promise.reject(error);
+      if (type !== "auth" && error?.response?.status === 403) {
+        encryptStorage.clear();
+        localStorage.clear();
+        window.location.href = "/admin";
       }
+      return Promise.reject(error);
     }
   );
